+++ app/src/main/java/com/courseproject/tindar/ui/conversation_list/placeholder/PlaceholderContent.java
@@ -1,4 +1,4 @@
-package com.courseproject.tindar.ui.conversations.placeholder;
+package com.courseproject.tindar.ui.conversation_list.placeholder;
 
 import java.util.ArrayList;
 import java.util.HashMap;
