+++ app/src/main/res/layout/fragment_conversation_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ui.conversations.ConversationFragment"
+    tools:context=".ui.conversation_list.ConversationFragment"
     tools:listitem="@layout/fragment_conversation" />
