+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/nav_conversations"
-        android:name="com.courseproject.tindar.ui.conversations.ConversationFragment"
+        android:name="com.courseproject.tindar.ui.conversation_list.ConversationFragment"
         android:label="Conversations"
         tools:layout="@layout/fragment_conversation" />
